+++ public/scripts/app.js
@@ -1,6 +1,13 @@
 "use strict";
 
 function _typeof(o) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, _typeof(o); }
+var _react = _interopRequireDefault(require("react"));
+var _reactDom = _interopRequireDefault(require("react-dom"));
+var _AddOption = _interopRequireDefault(require("../components/AddOption"));
+var _Option = _interopRequireDefault(require("../components/Option"));
+var _Header = _interopRequireDefault(require("../components/Header"));
+var _Action = _interopRequireDefault(require("../components/Action"));
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }
 function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, "prototype", { writable: false }); return Constructor; }
@@ -109,97 +116,37 @@ var Indecision = /*#__PURE__*/function (_React$Component) {
     value: function render() {
       var subTitle = 'Create short decision list to choose from';
       // const options = ['first', 'second']; // takes preference
-      return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement(Header, {
+      return /*#__PURE__*/_react["default"].createElement("div", null, /*#__PURE__*/_react["default"].createElement(_Header["default"], {
         subTitle: subTitle
-      }), /*#__PURE__*/React.createElement(Action, {
+      }), /*#__PURE__*/_react["default"].createElement(_Action["default"], {
         hasOptions: this.state.options.length > 0,
         makeDecision: this.makeDecision
-      }), /*#__PURE__*/React.createElement(Options, {
+      }), /*#__PURE__*/_react["default"].createElement(Options, {
         options: this.state.options,
         deleteOptions: this.deleteOptions,
         hasOptions: this.state.options.length > 0,
         deleteOption: this.deleteOption
-      }), /*#__PURE__*/React.createElement(AddOptions, {
+      }), /*#__PURE__*/_react["default"].createElement(AddOptions, {
         addOption: this.addOption
       }));
     }
   }]);
   return Indecision;
-}(React.Component); // Indecision.defaultProps = {
+}(_react["default"].Component); // Indecision.defaultProps = {
 //     options: []
 // }
-var Header = function Header(props) {
-  return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("h1", null, props.title), props.subTitle && /*#__PURE__*/React.createElement("h3", null, props.subTitle));
-};
-Header.defaultProps = {
-  title: 'Indecision App'
-};
-var Action = function Action(props) {
-  return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("button", {
-    disabled: !props.hasOptions,
-    onClick: props.makeDecision
-  }, "What should i do?"));
-};
-var AddOptions = /*#__PURE__*/function (_React$Component2) {
-  _inherits(AddOptions, _React$Component2);
-  var _super2 = _createSuper(AddOptions);
-  function AddOptions(props) {
-    var _this3;
-    _classCallCheck(this, AddOptions);
-    _this3 = _super2.call(this, props);
-    _this3.addOption = _this3.addOption.bind(_assertThisInitialized(_this3)); // binding the method
-    _this3.state = {
-      error: undefined
-    };
-    return _this3;
-  }
-  _createClass(AddOptions, [{
-    key: "addOption",
-    value: function addOption(e) {
-      e.preventDefault();
-      var option = e.target.elements.option.value.trim();
-      var error = this.props.addOption(option);
-      this.setState(function () {
-        return {
-          error: error
-        };
-      });
-      if (!error) {
-        e.target.elements.option.value = '';
-      }
-    }
-  }, {
-    key: "render",
-    value: function render() {
-      return /*#__PURE__*/React.createElement("div", null, this.state.error && /*#__PURE__*/React.createElement("p", null, this.state.error, " "), /*#__PURE__*/React.createElement("form", {
-        onSubmit: this.addOption
-      }, /*#__PURE__*/React.createElement("input", {
-        type: "text",
-        name: "option"
-      }), /*#__PURE__*/React.createElement("button", null, "Add Options")));
-    }
-  }]);
-  return AddOptions;
-}(React.Component);
 var Options = function Options(props) {
-  return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("button", {
+  return /*#__PURE__*/_react["default"].createElement("div", null, /*#__PURE__*/_react["default"].createElement("button", {
     disabled: !props.hasOptions,
     onClick: props.deleteOptions
-  }, "Remove All"), /*#__PURE__*/React.createElement("p", null, "You have ", props.options.length > 0 ? props.options.length : 'no', " options:"),
+  }, "Remove All"), /*#__PURE__*/_react["default"].createElement("p", null, "You have ", props.options.length > 0 ? props.options.length : 'no', " options:"),
   // this.props.options.map((option, index) => <p key={option}>{option} </p>)
   props.options.map(function (option, index) {
-    return /*#__PURE__*/React.createElement(Option, {
+    return /*#__PURE__*/_react["default"].createElement(_Option["default"], {
       key: option,
       optionText: option,
       deleteOption: props.deleteOption
     });
   }));
 };
-var Option = function Option(props) {
-  return /*#__PURE__*/React.createElement("div", null, props.optionText, /*#__PURE__*/React.createElement("button", {
-    onClick: function onClick(e) {
-      return props.deleteOption(props.optionText);
-    }
-  }, "Delete"));
-};
-ReactDOM.render( /*#__PURE__*/React.createElement(Indecision, null), document.getElementById('app'));
+_reactDom["default"].render( /*#__PURE__*/_react["default"].createElement(Indecision, null), document.getElementById('app'));
