+++ src/playground/stateless-indecision.js
@@ -1,8 +1,3 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import AddOption from '../components/AddOption'
-import Option from '../components/Option'
-
 class Indecision extends React.Component {
     constructor(props){
         super(props);
@@ -108,7 +103,37 @@ const Action = (props) => {
     )
 }
 
+class AddOptions extends React.Component {
+    constructor(props){
+        super(props);
+        this.addOption = this.addOption.bind(this); // binding the method
+        this.state = {
+            error: undefined
+        };
+    }
+    addOption(e){
+        e.preventDefault();
+        const option = e.target.elements.option.value.trim();
+        const error = this.props.addOption(option);
+        
+        this.setState(() => ({ error }));
 
+        if(!error) {
+            e.target.elements.option.value = '';
+        }
+    }
+    render() {
+        return (
+            <div>
+                { this.state.error && <p>{this.state.error} </p> }
+                <form onSubmit={this.addOption}>
+                    <input type="text" name="option" />
+                    <button>Add Options</button>
+                </form>
+            </div>
+        )
+    }
+}
 
 const Options = (props) => {
     return (
@@ -122,6 +147,15 @@ const Options = (props) => {
                 deleteOption={props.deleteOption} />
                 ))
             }
+        </div>
+    );
+};
+
+const Option = (props) => {
+    return (
+        <div>
+            {props.optionText}
+            <button onClick={(e) => props.deleteOption(props.optionText)}>Delete</button>
         </div>
     );
 };
